requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.3
django-cors-headers==4.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
fonttools==4.55.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kombu==5.4.2
lxml==5.3.0
msgpack==1.1.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.0
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
